backend/requirements.txt
python-dotenv
instructor
pydantic
groq
fastapi
uvicorn
google-generativeai